Modify default ResponderID in OCSP
The OCSP responses from OCSP responder changes the default from
ResponderID from Subject name to Key hash.

This change has been already done in the CA internal OCSP to limit
problems with repeated name in some complex scenarios.
